requirements.txt
nbtlib
multipledispatch
pillow
